New multi-integer version system (HaddonVersion)
I have decided that the previous system of using a single integer
to represent the version is insufficient when providing releases
for multiple versions of Minecraft.
Consider the following use case:
1. We create a release for 1.12.2, called 1.12.2-r20
2. We backport the release to 1.7.10, and call it 1.7.0-r20
3. We realize there is a bug in the backport which needs to be fixed.

The new release with the bugfix can't be called the same as the previous
version. The version number has to be incremented. So we would need to
call the bugfix version 1.7.10-r21.
But by doing this, we use up the version number of 21, so the next
release on the master branch (1.12.2) has to be called r22.
So, a bugfix on another branch incremented the version number on the
master branch.

If we instead used a traditional version format consisting of a sequence
of integers, this problem would be avoided. Bugfixes would not increment
the major version number.

---

One regression of this commit is that I removed version prefixes.
They were poorly implemented and didn't seem to be used anywhere.
However, they could easily be readded by changing the HaddonVersion
class.