Expand UpdateNotifier to update check external components
Instead of only checking the mod for updates, UpdateNotifier can
be given jobs to check UpdatableIdentity objects for updates
(upon being constructed, UpdateNotifier automatically creates a
UpdatableIdentity for the mod and adds it as its first job).
This way the UpdateNotifier can be used to check external components
of the mod such as resource packs for updates.

An UpdatableIdentity is an Identity plus a unique name and a
list of update urls.