Let UpdateNotifier decide platform version via a callback
A Haddon's "platform" refers to the Haddon's "parent".
For example, a Minecraft mod's platform is Minecraft, and its
platform version is Minecraft's version. A MAtmos soundpack's
platform is MAtmos, and its platform version is MAtmos's version.
This relationship is only a metaphor, there is no dependency
management or anything like that. In practice, this stuff is
only used for choosing which element of the update json's root
to choose updates from. So, in other words, the children of
the update json's root which are named as versions are defined
as platform versions.

When searching for new versions of a Haddon, it can vary which
entry of the JSON should be shown.

* When searching for new versions of a Minecraft mod, it makes
sense to only check updates for the version of Minecraft the mod
is running in. If someone is playing a modpack on 1.7.10 and the
mod keeps getting updates on 1.12.2, it is of no use to the player
to see that new versions keep coming out for 1.12.2, but if a new
version for 1.7.10 comes out, the player should be notified.

* When searching for new versions of a MAtmos soundpack, soundpack
versions shouldn't be shown if they were made for versions of the
MAtmos sound engine which are not available for the running version
of Minecraft. If the newest MAtmos version on 1.7.10 has a sound
engine version of 5, but 1.12.2 has one of 10, and the soundpack
keeps getting updates which require sound engine version 10, the
1.7.10 player should not be notified.

As you can see, the conditions for which platform versions to
consider can get quite varied. To let the UpdateNotifier decide
arbitrary conditions, a callback can be given as a parameter to
addJob(), which will be called with a set (actually a map) of
available platform versions to decide which platform version to
check for new updates.