If config file is missing, create blank config file
Prior to this commit, the config file would only be created
once the config has been modified, triggering a save.
So it required something like the user opening the MAtmos
menu (which saved the expansion volume levels), or getting
an update notification.

Now the config file is already created in the preinit
stage, which is useful if the game crashes before
even loading a world.

(Cherry pick of e9f0974a2322ba6a4c75dd7c1b8fb9feab09370d )